+++ src/cuda/sigmoid_pairwise_kernels.cu
@@ -2,7 +2,7 @@
 #include <cstddef>
 #include "constants.cuh"
 
-__global__ void sigmoid_pairwise_loss_kernel(int32_t* y_true, float* exp_pred, float* sigmoid_loss, size_t N){
+__global__ void sigmoid_pairwise_loss_kernel(int32_t* y_true, float* exp_pred, float* sigmoid_loss, uint N){
     int gid = threadIdx.x + blockDim.x*blockIdx.x;
 
     __shared__ float _loss;
@@ -13,11 +13,11 @@ __global__ void sigmoid_pairwise_loss_kernel(int32_t* y_true, float* exp_pred, f
     __syncthreads();
 
     while(gid < N){
-        size_t _i = (gid%2 == 1) ? (N - size_t(gid/2) - 1) : size_t(gid/2);
-        for(size_t j = 0; j <= _i; j++){
+        uint _i = (gid%2 == 1) ? (N - uint(gid/2) - 1) : uint(gid/2);
+        for(uint j = 0; j <= _i; j++){
             float P_hat = 0.5f*(y_true[_i] - y_true[j]) + 0.5f;
             float P = 1.f / (1.f + (exp_pred[j] / exp_pred[_i]));
-            float delta_loss = P_hat*log(P + EPS_CU) + (1.f - P_hat)*log(1.f - P - EPS_CU);
+            float delta_loss = P_hat*logf(P + EPS_CU) + (1.f - P_hat)*logf(1.f - P - EPS_CU);
             atomicAdd(&_loss, delta_loss);
         }
         gid += blockDim.x*gridDim.x;
@@ -29,7 +29,7 @@ __global__ void sigmoid_pairwise_loss_kernel(int32_t* y_true, float* exp_pred, f
     }
 }
 
-__global__ void sigmoid_pairwise_grad_hess_kernel(int32_t* y_true, float* exp_pred, float* grad, float* hess, size_t N){
+__global__ void sigmoid_pairwise_grad_hess_kernel(int32_t* y_true, float* exp_pred, float* grad, float* hess, uint N){
     int gid = threadIdx.x + blockDim.x*blockIdx.x;
 
     float exp_tmp_diff = 0.;
@@ -41,8 +41,8 @@ __global__ void sigmoid_pairwise_grad_hess_kernel(int32_t* y_true, float* exp_pr
     float P_hat = 0.;
 
     while(gid < N){
-        size_t _i = (gid%2 == 1) ? (N - size_t(gid/2) - 1) : size_t(gid/2);
-        for(size_t j = 0; j < _i; j++){
+        uint _i = (gid%2 == 1) ? (N - uint(gid/2) - 1) : uint(gid/2);
+        for(uint j = 0; j < _i; j++){
             P_hat = 0.5f*(y_true[_i] - y_true[j]) + 0.5f;
             exp_tmp_diff = exp_pred[_i] / exp_pred[j];
